Add files via upload
This repository includes the second test we ran on our robot. In this code, we have added and made progress on the previous line following code. Here, we've used a NavX magnetometer, in which we have tried to limit the number of turns out line following robot takes based on the values detected and stored by the Nav X sensor. In a later repository we will add more information on the sensor we have used and how we are implementing the sensor's values.